+++ CHANGELOG.rst
@@ -15,6 +15,7 @@ Added
 - Handled ``kytos/mef_eline.uni_active_updated`` event.
 - Handled ``kytos/mef_eline.deployed`` event.
 - Handled ``kytos/mef_eline.(failover_link_down|failover_old_path|failover_deployed)`` events.
+- Added UI for telemetry_int.
 
 Changed
 =======